Decentralize all samples
With the idea that we want to roll a lot more samples for Gradle and
Nokee, we need a way to add and remove them easily. Part of that
process, we want the samples to be easily developped meaning that we can
actually play with them without requiring a standalone, unidirectional,
copy. For this purpose, the work in this commit is to have the sample
project be "a sample" which essentially contains all of the plumbing to
create the end samples (and the variants). The new sample infrastructure
will focus on a single DSL which we will transform into the other DSL.
It will allow to standardize the whole process and create high quality
samples. Upon generating the sample, we will remove all of the sample
plumbing to create a "pure" sample. We will release those "pure sample"
in GitHub repos or our sample portal.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>