Introduce "producing/consuming" samples
The idea is to create an artifact that will represent a sample which we
can then transform into other artifact for the various needs. For
example, if we bake the samples using JBake, we may want to transform
into JBake content/assets. We may also want to create an sample index.
Both of which are not the responsability of the sample.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>