Allow samples to be imported as external Gradle projects
This allow to work with each individual samples directly from IntelliJ.
There are a couple trick that we need to pull off in order to make this
work. First we need to inspect each sample to list the composite builds
and their modules. Then, we need to compute a look aside table of sample
and their name so we can override the project name via init script so
they appear correctly in IntelliJ. Finally, we have to write the
.idea/gradle.xml file to include all external projects (and the current
one). One downside with this approach is the inability to perform full
sync without issues. IntelliJ sync is too short sigthed that it will for
as many Gradle instance as they are external project quickly overloading
the computer.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>