Introduce very early rendering of all samples
The goal is to allow the main sample repository to consume this
repository (and other sample repositories) to render the whole sample
site. It's important to keep a distinction between the sample and the
rendering. This is why we export a sample package which include a
manifest and the sample variants as its content. From the sample
package, we should be able to derived the required artifact for all
samples.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>